Improve type safety of transformer factory
Each transformer now holds a type token which is used by the factory to
find the right transformer. But because the class literal for primitive
types returns another class object than the literal for the wrapper
types (i.e. long.class is another object than Long.class), there needs
to be a transformer for each primitive type as well.